requirements.txt
asgiref==3.3.1
Django==3.1.6
django-leaflet==0.27.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
